+++ test/test-cjs.cjs
@@ -1,6 +1,3 @@
-// TODO:
-// 1. Test Sequential Calls to the Same Browser
-
 const { html2pdf, img2pdf, pdf2img, launchBrowser } = require ( "../dist/index.cjs" )
 const { mkdirSync, statSync, readFileSync } = require ( "fs" )
 const { rm, writeFile } = require ( "fs/promises" )
@@ -24,7 +21,8 @@ if ( !isDirectory ( testAssetsDir ) ) {
 
 const runTests = async ( ) => {
   try {
-    await launchBrowser ( "firefox", "ws://172.21.100.71:3000/devtools/browser/307fe387-60f6-41ed-b1ab-eb1d27400077" )
+    // Use browserless/chrome for testing
+    // await launchBrowser ( "chrome", "ws:localhost:3000" )
     const pdf = "https://www.orimi.com/pdf-test.pdf"
 
     const pdf2imgResult = await pdf2img ( pdf )
@@ -46,7 +44,7 @@ const runTests = async ( ) => {
     await writeFile ( resolve ( testAssetsDir, "html-to-pdf.pdf" ), html )
     console.log ( "HTML to PDF Conversion Successful (CommonJS)" )
   } catch ( error ) {
-    console.error ( "Error during CommonJS tests:", JSON.stringify ( error, null, 2 ) )
+    console.error ( "An error occurred during the tests:", error )
   } finally {
     // Clean up the test-assets directory after the tests
     await rm ( testAssetsDir, { force: true, recursive: true } )